lib: Add detiling routines
As determined by marcan, it's a standard Z-order interleave with 64-bit
tiles. Add some basic detiling routines for 32bpp, other forms will
generalize easily like we do in the Panfrost code.

It's not really optimal but it's faster than the obvious thing and it's
certainly fast enough. Oddly NEON is making things slower on RK3399,
will need to test on the M1.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>