disasm: Add store vary opcode
Stores a single 32-bit scalar. Note that I/O is _not_ vectorized, which
admittedly makes for easier register allocation and somewhat more
"interesting" schedules. Example:

	11 10 82 80

This is storing the 32-bit register (0x10 >> 1) = w8 to varying slot 2
(indicated in the lower nibble of the third byte). That covers the first
byte, 7 bits of the second byte, and ~4 bits of the third byte. The
meaning of the remaining ~13 bits is not yet clear.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>